[Design] #5 - Cell에서 HeaderView로 변경
