--- every-time-iOS/every-time-iOS/Global/Extension/CGFloat +.swift
@@ -1,17 +1,0 @@
-//
-//  CGFloat +.swift
-//  every-time-iOS
-//
-//  Created by 김민 on 2023/01/16.
-//
-
-import UIKit
-
-extension CGFloat {
-    
-    var adjusted: CGFloat {
-        let ratio: CGFloat = UIScreen.main.bounds.width / 375
-        let ratioH: CGFloat = UIScreen.main.bounds.height / 667
-        return ratio <= ratioH ? self * ratio : self * ratioH
-    }
-}
