+++ every-time-iOS/every-time-iOS/Global/Extension/UITableViewHeaderFooterView +.swift
@@ -1,0 +1,15 @@
+//
+//  UITableViewHeaderFooterView +.swift
+//  every-time-iOS
+//
+//  Created by 김민 on 2023/01/24.
+//
+
+import UIKit
+
+extension UITableViewHeaderFooterView {
+    
+    static var cellIdentifier: String {
+        return String(describing: self)
+    }
+}
