--- every-time-iOS/every-time-iOS/Presentation/Home/Cell/TableViewCell/BannerTableViewCell.swift
@@ -1,59 +1,0 @@
-//
-//  BannerTableViewCell.swift
-//  every-time-iOS
-//
-//  Created by 김민 on 2023/01/23.
-//
-
-import UIKit
-
-import SnapKit
-import Then
-
-final class BannerTableViewCell: UITableViewCell {
-    
-    // MARK: - UI Components
-    
-    private let bannerImageView: UIImageView = UIImageView()
-    
-    // MARK: - Initializer
-    
-    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-        
-        setUI()
-        setLayout()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
-extension BannerTableViewCell {
-    
-    // MARK: - UI Components Property
-    
-    private func setUI() {
-        selectionStyle = .none
-        
-        contentView.backgroundColor = .clear
-        
-        bannerImageView.do {
-            $0.backgroundColor = .systemGray4
-            $0.layer.cornerRadius = 7
-        }
-    }
-    
-    // MARK: - Layout Helper
-    
-    private func setLayout() {
-        contentView.addSubview(bannerImageView)
-        
-        bannerImageView.snp.makeConstraints {
-            $0.top.bottom.equalToSuperview()
-            $0.leading.equalToSuperview().offset(10)
-            $0.trailing.equalToSuperview().offset(-10)
-        }
-    }
-}
