--- every-time-iOS/every-time-iOS/Presentation/Home/Cell/TableViewCell/MenuTableViewCell.swift
@@ -1,106 +1,0 @@
-//
-//  MenuTableViewCell.swift
-//  every-time-iOS
-//
-//  Created by 김민 on 2023/01/23.
-//
-
-import UIKit
-
-import SnapKit
-import Then
-
-final class MenuTableViewCell: UITableViewCell {
-    
-    // MARK: - UI Components
-    
-    private let menuCollectionView: UICollectionView = {
-        let layout = UICollectionViewFlowLayout()
-        layout.scrollDirection = .horizontal
-        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
-        return collectionView
-    }()
-    
-    // MARK: - Initializer
-    
-    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-        
-        setUI()
-        setLayout()
-        setDelegate()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
-extension MenuTableViewCell {
-    
-    // MARK: - UI Components Property
-    
-    private func setUI() {
-        selectionStyle = .none
-        
-        contentView.backgroundColor = .clear
-        
-        menuCollectionView.do {
-            $0.backgroundColor = .clear
-            $0.translatesAutoresizingMaskIntoConstraints = false
-            $0.showsHorizontalScrollIndicator = false
-            $0.register(MenuCollectionViewCell.self, forCellWithReuseIdentifier: MenuCollectionViewCell.cellIdentifier)
-        }
-    }
-    
-    // MARK: - Layout Helper
-    
-    private func setLayout() {
-        contentView.addSubview(menuCollectionView)
-        
-        menuCollectionView.snp.makeConstraints {
-            $0.edges.equalToSuperview()
-        }
-    }
-    
-    // MARK: - Methods
-    
-    private func setDelegate() {
-        menuCollectionView.dataSource = self
-        menuCollectionView.delegate = self
-    }
-}
-
-// MARK: - UICollectionViewDataSource
-
-extension MenuTableViewCell: UICollectionViewDataSource {
-    
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return 7
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: MenuCollectionViewCell.cellIdentifier, for: indexPath)
-        return cell
-    }
-}
-
-// MARK: - UICollectionViewDelegateFlowLayout
-
-extension MenuTableViewCell: UICollectionViewDelegateFlowLayout {
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumLineSpacingForSectionAt section: Int) -> CGFloat {
-        return 7
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, insetForSectionAt section: Int) -> UIEdgeInsets {
-        return UIEdgeInsets(top: 0, left: 20, bottom: 0, right: 20)
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        let screenWidth = UIScreen.main.bounds.width
-        let cellWidth = screenWidth - 20 - 7 * 6
-        let height = contentView.frame.height - 20
-        return CGSize(width: cellWidth/7, height: height)
-    }
-}
